fix bug in expire cache

+++ lib/expire_cache.rb
@@ -4,10 +4,12 @@ module ExpireCache
 
   private
     def uncache sub_path
-      path = "#{ExpireCache::CACHE_ROOT}#{sub_path}"
-      if File.exist?(path)
-        puts 'deleting: ' + sub_path
-        File.delete(path)
+      if sub_path
+        path = "#{ExpireCache::CACHE_ROOT}#{sub_path}"
+        if File.exist?(path)
+          puts 'deleting: ' + sub_path
+          File.delete(path)
+        end
       end
     end
